Hello... it's been 3? days since the last commit. What i have got done in these few days has been a decent amount of stuff. I had some issues with finding the PCIe ports properly for some reason since my search wasnt finding the NVME for some reason. Then i checked on my laptop and it found that. Then i re-checked on laptop and it didnt find it again... I dont know why that happened but it should be fixed. (At least i find my nvme on both PC and laptop now...) Also fixed one issue where on my real PC the Millis went by like 400x faster than they shouldve. but everything should be now resolved. I feel like i got less done in these last 3 days than in the first 2 days (Propably because i lowk spent the whole yesterday playing Apex Legends cause new season). Propably the next goal is to make some types of commands so that i can create the partitions and find empty space on the disk. -'The silicon was never silet. Only unadressed' -Sun Tzu, The Art of the Adress Space

+++ Makefile
@@ -31,13 +31,17 @@ $(BUILD_DIR)/%.o: src/%.cc $(HEADERS)
 run: BOOTX64.EFI
 	mkdir -p esp/EFI/BOOT
 	cp BOOTX64.EFI esp/EFI/BOOT/
-	cp $(OVMF_VARS) ./OVMF_VARS.fd
+	[ -f OVMF_VARS.fd ] || cp $(OVMF_VARS) ./OVMF_VARS.fd
+	[ -f disk.img ] || qemu-img create -f raw disk.img 64M
+
 	qemu-system-x86_64 \
 		-m 1G \
 		-drive if=pflash,format=raw,unit=0,file=$(OVMF_CODE),readonly=on \
 		-drive if=pflash,format=raw,unit=1,file=OVMF_VARS.fd \
 		-drive format=raw,file=fat:rw:esp \
-		-serial stdio
+		-serial stdio \
+		-drive file=disk.img,if=none,id=nvm,format=raw \
+		-device nvme,serial=deadbeed,drive=nvm \
 
 clean:
 	rm -rf BOOTX64.EFI esp OVMF_VARS.fd $(BUILD_DIR)
